+++ src/server/headless.rs
@@ -7139,6 +7139,7 @@ next_tab = ""
                 ahead_behind: None,
                 space: None,
                 change_set: None,
+                collectible: None,
             }],
             cache_updates: Vec::new(),
         });
@@ -7166,6 +7167,7 @@ next_tab = ""
                 ahead_behind: None,
                 space: None,
                 change_set: None,
+                collectible: None,
             }],
             cache_updates: Vec::new(),
         });